Add and implement BaseApplication
* Introduces BaseApplication to reduce duplication across apps
* Consolidates shared initialization, routing, and middleware logic
* Simplifies V1/V2 applications through inheritance
* Standardizes Rack app configuration and middleware stack
* Improves maintainability by centralizing common application code
